+++ src/components/about/edu.tsx
