+++ src/components/blogs/listpage.tsx
@@ -1,0 +1,11 @@
+import { BlogsPageHeaderGrp } from "./header";
+import { BlogsContent } from "./contents";
+
+export default function BlogsPage() {
+    return (
+        <>
+            <BlogsPageHeaderGrp />
+            <BlogsContent />
+        </>
+    );
+}
