+++ src/components/layouts/main.tsx
@@ -1,0 +1,13 @@
+import Navbar from "../navbar";
+import { Outlet } from "react-router-dom";
+
+function MainLayout() {
+    return (
+        <>
+            <Navbar />
+            <Outlet />
+        </>
+    );
+}
+
+export default MainLayout;
