+++ src/components/projects/header.tsx
@@ -1,0 +1,13 @@
+export default function ProjectsPageHeaderGrp() {
+    return (
+      <div className='header-grp'>
+        <div className='header-txt'>
+          <b>Projects</b>
+        </div>
+        <div className='description-txt mt-2'>
+          I just have studied Information Technology for two and a half years, {" "}
+          but luckily have some fun works done.
+        </div>
+      </div>
+    );
+}
