+++ src/components/projects/page.tsx
@@ -1,17 +1,18 @@
 import { motion as m } from "framer-motion";
-import GuestbookHeader from "./header";
+import ProjectsPageHeaderGrp from "./header";
+import { ProjectsList } from "./projects";
 import Copyright from "../copyright";
-import GuestbookBody from "./body";
 
-
-export default function GuestbookPage() {
-    return(
+export default function ProjectsPage() {
+    return (
         <m.div
             transition={{ duration: 0.8 }}
             exit={{ opacity: 0 }} 
             className="container page-wrapper transition-page">
-            <GuestbookHeader />
-            <GuestbookBody />
+            <ProjectsPageHeaderGrp />
+            <div className="project-wrapper">
+                <ProjectsList />
+            </div>
             <Copyright copyright_class="page-copyright pb-3" />
         </m.div>
     );
